requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
influxdb-client==1.36.1
psycopg2-binary==2.9.9
boto3==1.34.34
python-multipart==0.0.9
pydantic==2.6.1 